requirements.txt
# Core Dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2

# LLM & Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langgraph==0.0.20
openai==1.6.1
google-generativeai==0.3.2

# Data & Query Engine
duckdb==0.9.2
pyarrow==14.0.1

# Vector Store (Optional but included for scalability)
faiss-cpu==1.7.4
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Visualization & UI
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
